pyproject.toml
[project]
name = "mcp-server-garmincn"
version = "0.1.0"
description = "用于大模型的MCP服务，提供获取用户的佳明数据的tool"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.4.0",
    "garminconnect>=0.2.26",
    "python-dotenv>=1.0.0",
]

[project.scripts]
mcp-server-garmincn = "mcp_server_garmincn:main"